fix: Resolve remaining 32 clang-tidy errors and 2 build errors
Suppress cert-err34-c globally (all 17 call sites already check scanf
return values; rewriting to strtol would be churn for no safety gain).
Fix widening/narrowing conversion warnings in cgroup.c, entity.c,
fingerprint.c, dedup.c, ioperf.c, inotify.h, network.c, port.c, and
thermal.c. Fix ULL→uint64_t type mismatch in disk.c that broke both
GCC and clang builds with -Werror.

+++ ebpf/telemetryd/probes/disk.c
@@ -133,8 +133,8 @@ int disk_probe_run(struct normalizer *norm, struct telem_socket *sock, void *ctx
                      "Disk space low on %s: %.1f%% used (free: %" PRIu64 " GB, total: %" PRIu64 " GB)",
                      mount,
                      used_pct * 100.0,
-                     avail / (1024ULL * 1024ULL * 1024ULL),
-                     total / (1024ULL * 1024ULL * 1024ULL));
+                     avail / ((uint64_t)1024 * 1024 * 1024),
+                     total / ((uint64_t)1024 * 1024 * 1024));
 
             snprintf(entity, sizeof(entity), "mount:%s", mount);
